+++ formula-android/src/main/java/com/instacart/formula/android/internal/FragmentFlowRenderView.kt
@@ -18,7 +18,7 @@ import com.instacart.formula.fragment.FragmentLifecycleEvent
 import com.instacart.formula.fragment.getFormulaFragmentId
 import com.instacart.formula.integration.BackCallback
 import com.instacart.formula.integration.FeatureEvent
-import com.instacart.formula.integration.FragmentId
+import com.instacart.formula.android.FragmentId
 import java.util.LinkedList
 import java.util.UUID